Give Flyway ownership of the schema and reconcile entity mappings
The repository carried three conflicting definitions of the same schema -
tableSchema.sql, a set of per-table .sql files, and the JPA entities -
and all three disagreed. Because ddl-auto was disabled, nothing ever
checked them against each other, so the drift accumulated silently.

The drift was not cosmetic. Trade and Position mapped the literal column
"stockTicker" while every SQL file declared stock_ticker, so those
entities could not have queried the documented schema at all. Order,
Trade and Options each joined Stock via referencedColumnName
"stockTicker", a column that does not exist - Stock's id column is
stock_ticker. Stock mapped number_of_share against a number_of_shares
column. The prod profile already set ddl-auto=validate, so production
could not have started.

Two of those broken joins were introduced by me: while fixing an
unrelated @JoinColumn defect in the authorization change, I pointed
Order.stock at referencedColumnName "stockTicker". Flagging it here
rather than repairing it quietly inside other work.

Reconciliation:
- Every column mapping standardised on snake_case, matching Stock's id
  column and the SQL files. This covers @Column names, @JoinColumn names
  and referencedColumnName values, @Index columnLists (which also
  referenced executionDate), the watchlist join table's inverse join
  column, and Stock's element-collection join column.
- Order.stockTicker, previously relying on the implicit naming strategy,
  is now mapped explicitly.

Flyway:
- V1__baseline_schema.sql is authored from the corrected entities and
  becomes the sole source of schema truth. Version comes from Boot 2.6.6
  (8.0.5), where MySQL support still ships inside flyway-core; the split
  flyway-mysql artifact does not exist before 8.2.1.
- ddl-auto=validate moves into the base configuration rather than prod
  only, so drift now fails at startup in dev instead of surfacing later
  as a runtime query error.
- The legacy SQL files move to docs/legacy-schema/ with a README
  recording what each contained and how they contradicted one another.
  They are kept because they describe the shape of any older database
  dump, not because anything should apply them.

Money precision, applying the boundary rule in CLAUDE.md - a value is
money once multiplied by a quantity, settled, or persisted to a position,
ledger or trade:
- Converted to BigDecimal: Position.unrealizedPnL (P&L is money by
  invariant, and its column was already DECIMAL(19,4)), Order.targetPrice
  (a limit price is an order value), and OptionTradeDetails.strikePrice
  (which also makes it consistent with Options.strikePrice, already
  BigDecimal).
- Deliberately left as double: Stock's ask, bid, open, close, last,
  markChange, marketCapAmount and sizzleIndex. These are raw market data
  that is displayed rather than settled. Converting them would be
  applying the rule by reflex; the multiplication that turns a quote into
  money already happens in BigDecimal inside TradeServiceImpl.
- Quantities stay Double. Fractional-share precision is a separate
  question and out of scope here.

The migration itself cannot be executed in this environment - there is no
Docker and no MySQL - so V1 is verified by reading it against each
mapping rather than by applying it. Booting against a real MySQL is the
authoritative check and is noted in the pull request.

Verified with mvn clean compile.

+++ src/main/java/com/moneyteam/trading/model/OptionTradeDetails.java
@@ -1,5 +1,6 @@
 package com.moneyteam.trading.model;
 
+import java.math.BigDecimal;
 import javax.persistence.*;
 import java.time.LocalDate;
 
@@ -11,7 +12,7 @@ public class OptionTradeDetails {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    private Double strikePrice;
+    private BigDecimal strikePrice;
     private LocalDate expirationDate;
     private String optionType; // CALL or PUT
 
@@ -29,11 +30,11 @@ public class OptionTradeDetails {
         this.id = id;
     }
 
-    public Double getStrikePrice() {
+    public BigDecimal getStrikePrice() {
         return strikePrice;
     }
 
-    public void setStrikePrice(Double strikePrice) {
+    public void setStrikePrice(BigDecimal strikePrice) {
         this.strikePrice = strikePrice;
     }
